Fix some smaller type inconsistencies revealed by the last commit.

+++ hydpy/interfaces/routinginterfaces.py
@@ -17,7 +17,7 @@ class CrossSectionModel_V1(modeltools.SubmodelInterface):
     typeid: ClassVar[Literal[1]] = 1
     """Type identifier for |CrossSectionModel_V1| submodels."""
 
-    def prepare_bottomslope(self, bottomslope: int) -> None:
+    def prepare_bottomslope(self, bottomslope: float) -> None:
         """Set the bottom's slope (in the longitudinal direction) [-]."""
 
     def get_depths_of_discontinuity(self) -> tuple[float, ...]:
